M3T1 Refactoring
