+++ src/const.js
@@ -1,0 +1,1 @@
+export const COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
