+++ src/view/task-edit.js
@@ -1,18 +1,6 @@
-const isExpired = (dueDate) => {
-  if (dueDate === null) {
-    return false;
-  }
+import {COLORS} from "../const.js";
 
-  const currentDate = new Date();
-
-  currentDate.setHours(23, 59, 59, 999);
-
-  return currentDate > dueDate.getTime();
-};
-
-const isRepeating = (repeating) => {
-  return Object.values(repeating).some(Boolean);
-};
+import {isTaskExpired, isTaskRepeating, humanizeTaskDueDate} from "../utils.js";
 
 const createTaskEditorDateTemplate = (dueDate) => {
   return `<button class="card__date-deadline-toggle" type="button">
@@ -25,7 +13,7 @@ const createTaskEditorDateTemplate = (dueDate) => {
           type="text"
           placeholder=""
           name="date"
-          value="${dueDate.toLocaleString(`en-US`, {day: `numeric`, month: `long`})}"
+          value="${humanizeTaskDueDate(dueDate)}"
         />
       </label>
     </fieldset>` : ``}
@@ -34,9 +22,9 @@ const createTaskEditorDateTemplate = (dueDate) => {
 
 const createTaskEditorRepeatingTemplate = (repeating) => {
   return `<button class="card__repeat-toggle" type="button">
-    repeat:<span class="card__repeat-status">${isRepeating(repeating) ? `yes` : `no`}</span>
+    repeat:<span class="card__repeat-status">${isTaskRepeating(repeating) ? `yes` : `no`}</span>
   </button>
-  ${isRepeating(repeating) ? `<fieldset class="card__repeat-days">
+  ${isTaskRepeating(repeating) ? `<fieldset class="card__repeat-days">
     <div class="card__repeat-days-inner">
       ${Object.entries(repeating).map(([day, repeat]) => `<input
         class="visually-hidden card__repeat-day-input"
@@ -54,9 +42,7 @@ const createTaskEditorRepeatingTemplate = (repeating) => {
 };
 
 const createTaskEditorColorsTemplate = (currentColor) => {
-  const colors = [`black`, `yellow`, `blue`, `green`, `pink`];
-
-  return colors.map((color) => `<input
+  return COLORS.map((color) => `<input
     type="radio"
     id="color-${color}"
     class="card__color-input card__color-input--${color} visually-hidden"
@@ -87,12 +73,12 @@ export const createTaskEditorTemplate = (task = {}) => {
     }
   } = task;
 
-  const deadlineClassName = isExpired(dueDate) ?
+  const deadlineClassName = isTaskExpired(dueDate) ?
     `card--deadline` :
     ``;
   const dateTemplate = createTaskEditorDateTemplate(dueDate);
 
-  const repeatingClassName = isRepeating(repeating)
+  const repeatingClassName = isTaskRepeating(repeating)
     ? `card--repeat`
     : ``;
   const repeatingTemplate = createTaskEditorRepeatingTemplate(repeating);
