+++ src/view/task.js
@@ -1,31 +1,17 @@
-const isExpired = (dueDate) => {
-  if (dueDate === null) {
-    return false;
-  }
-
-  let currentDate = new Date();
-  currentDate.setHours(23, 59, 59, 999);
-  currentDate = new Date(currentDate);
-
-  return currentDate.getTime() > dueDate.getTime();
-};
-
-const isRepeating = (repeating) => {
-  return Object.values(repeating).some(Boolean);
-};
+import {isTaskExpired, isTaskRepeating, humanizeTaskDueDate} from "../utils";
 
 export const createTaskTemplate = (task) => {
   const {color, description, dueDate, repeating, isArchive, isFavorite} = task;
 
   const date = dueDate !== null ?
-    dueDate.toLocaleString(`en-US`, {day: `numeric`, month: `long`}) :
+    humanizeTaskDueDate(dueDate) :
     ``;
 
-  const deadlineClassName = isExpired(dueDate) ?
+  const deadlineClassName = isTaskExpired(dueDate) ?
     `card--deadline` :
     ``;
 
-  const repeatClassName = isRepeating(repeating) ?
+  const repeatClassName = isTaskRepeating(repeating) ?
     `card--repeat` :
     ``;
